Add files via upload
It is a tool to hack passwords of the networks you are connected to now. This means that if I am connected to a network called Network, and I run the tool and type the name of the network that I am now connected to, it will show me the password for the network. It can be tried in cafes and places where you always have to pay money to enjoy the Internet, where you can pay once to take the network, then run the tool and write the name of the network, then you will have the password with you easily.