Update README.md
It is a tool for hacking the networks you are connected to to show the password of the network you are connected to now, for example, if you are connected to a café network and you have to pay money to enjoy the Internet, then you paid today and connected to the network and saw that the name of the network is, for example (Cafe 3333), the tool's function comes from here. You run the tool, write the name of the network you are connected to, then the tool will show you the network password, which leads to you getting the Internet service for free without paying anything after today. If you do not remember the password of your network, but you are connected to it, run the tool and write the name of the network you are connected to now, its password will appear to you without going to the company to change it.